Added top companies and genres
