+++ src/hooks/useGenres.ts
@@ -1,7 +1,7 @@
 import config from '../api-clients/endpoints'
 import { getRequestConfig } from './request/requestConfig'
 import baseRequest from './request/baseRequest'
-import { IGenre } from '../types/Game'
+import { IGenre } from '../types/Genre'
 
 const useGenres = () => {
   const endpoint = `${config.gamesSvc.domain}${config.gamesSvc.genres}`
@@ -12,8 +12,15 @@ const useGenres = () => {
     return response
   }
 
+  const fetchTopGenres = async () => {
+    const url = `${endpoint}/top`
+    const response = await baseRequest<IGenre[]>(url, getRequestConfig)
+    return response
+  }
+
   return {
-    fetchGenres
+    fetchGenres,
+    fetchTopGenres
   }
 }
