+++ src/types/Company.ts
@@ -1,0 +1,9 @@
+/**
+ * ICompany represents company model
+ */
+interface ICompany {
+	id: number,
+	name: string
+}
+
+export type { ICompany }
