+++ src/types/Genre.ts
@@ -1,0 +1,9 @@
+/**
+ * IGenre represents genre model
+ */
+interface IGenre {
+	id: number,
+	name: string
+}
+
+export type { IGenre }
