+++ src/types/Platform.ts
@@ -1,0 +1,10 @@
+/**
+ * IPlatform represents platform model
+ */
+interface IPlatform {
+	id: number,
+	name: string,
+	abbreviation: string
+}
+
+export type { IPlatform }
