+++ src/utils/validators.ts
@@ -1,0 +1,8 @@
+export const IsValidUrl = (url: string): boolean => {
+  try {
+    new URL(url)
+    return true
+  } catch (err) {
+    return false
+  }
+}
