+++ src/views/Landing.tsx
@@ -1,37 +1,39 @@
 import React, { useState, useEffect } from 'react'
 import {
-  Alert, AlertColor, Autocomplete, AutocompleteChangeReason, AutocompleteChangeDetails, Backdrop, Box, Button, 
-  CircularProgress, Container, Grid, Pagination, Snackbar, SnackbarCloseReason, Stack, TextField, TextFieldProps, 
-  Typography, ToggleButton, ToggleButtonGroup, useMediaQuery, useTheme
+  Backdrop, Box, Button, CircularProgress, Container, Grid, Pagination, Stack, Typography, 
+  ToggleButton, ToggleButtonGroup, useMediaQuery, useTheme, Chip, ListItem, List, ListItemText, ListItemButton
 } from '@mui/material'
 import { AdapterMoment as DateAdapter } from '@mui/x-date-pickers/AdapterMoment'
-import { MobileDatePicker, DesktopDatePicker, LocalizationProvider } from '@mui/x-date-pickers'
+import { LocalizationProvider } from '@mui/x-date-pickers'
 import AbcIcon from '@mui/icons-material/AbcRounded'
 import WhatshotIcon from '@mui/icons-material/WhatshotRounded'
 import DateRangeIcon from '@mui/icons-material/DateRangeRounded'
-import moment from 'moment'
-import { FileInfo, FileUpload, FilesUpload, Widget as UploadWidget } from '@uploadcare/react-widget'
+import CategoryOutlinedIcon from '@mui/icons-material/CategoryOutlined'
+import KeyboardArrowDownOutlinedIcon from '@mui/icons-material/KeyboardArrowDownOutlined'
+import KeyboardArrowUpOutlinedIcon from '@mui/icons-material/KeyboardArrowUpOutlined';
 import { useSearchParams } from 'react-router-dom'
 
 import Layout from '../components/Layout'
 import GameCard from '../components/GameCard'
 import GameDetails from '../components/GameDetails'
+import AddGameModal from '../components/AddGameModal'
+import Notification from '../components/Notification'
 import { IDarkModeProps, ISearchFieldProps } from '../components/Header'
-import { ICreateGame, IGame, IGames, IGamesFilter, IGameResponse, IPlatform, IGenre } from '../types/Game'
+import { IGame, IGames, IGamesFilter } from '../types/Game'
+import { IGenre } from '../types/Genre'
+import { ICompany } from '../types/Company'
 import useGames from '../hooks/useGames'
 import useGenres from '../hooks/useGenres'
-import usePlatforms from '../hooks/usePlatforms'
+import useCompanies from '../hooks/useCompanies'
 import useUser from '../hooks/useUser'
 import useAuth from '../hooks/useAuth'
 import { IValidationResponse } from '../types/Validation'
 import { roles } from '../auth/roles'
 import { IGetUserRatingsResponse } from '../types/Rating'
-import { isTouchDevice } from '../utils/devices'
-import Modal from '../components/Modal'
 import '../styles/UploadWidget.css'
 
-const fieldWidthLarge = '400px'
-const fieldWidthSmall = '210px'
+const topCategoriesLimit = 8
+const selectedCategoryColor = "#1ea0c0"
 
 interface ILandingProps {
   darkModeProps: IDarkModeProps
@@ -40,22 +42,34 @@ interface ILandingProps {
 const Landing = (props: ILandingProps) => {
   const { darkModeProps } = props
 
-  const { fetchPage: fetchGames, create: createGame } = useGames()
+  const { fetchPage: fetchGames } = useGames()
   const { fetchRatings } = useUser()
-  const { fetchGenres } = useGenres()
-  const { fetchPlatforms } = usePlatforms()
+  const { fetchTopGenres } = useGenres()
+  const { fetchTopCompanies } = useCompanies()
   const { hasRole, isAuthenticated } = useAuth()
+
   const theme = useTheme()
-  const matchesMd = useMediaQuery(theme.breakpoints.up('md'))
   const matchesSm = useMediaQuery(theme.breakpoints.only('sm'))
   const matchesXs = useMediaQuery(theme.breakpoints.only('xs'))
 
   const [data, setData] = useState<IGame[]>([])
   const [count, setCount] = useState<number>(0)
   const [userRatings, setUserRatings] = useState<IGetUserRatingsResponse>({})
-  const [genres, setGenres] = useState<IGenre[]>([])
-  const [platforms, setPlatforms] = useState<IPlatform[]>([])
+  const [topGenres, setTopGenres] = useState<IGenre[]>([])
+  const [topDevelopers, setTopDevelopers] = useState<ICompany[]>([])
+  const [topPublishers, setTopPublishers] = useState<ICompany[]>([])
+  const [topCategoriesOpen, setTopCategoriesOpen] = useState<boolean>(false)
   const [isLoading, setIsLoading] = useState(false)
+
+  const [alert, setAlert] = useState<string | IValidationResponse | null>(null)
+  
+  const mediaQueryToSize = (): "small" | "medium" | "large" => {
+    return matchesXs ? "small" : matchesSm ? "medium" : "large"
+  }
+
+  const mediaQueryToTextSize = (): "small" | "medium" | "large" => {
+    return matchesXs ? "small" :"medium"
+  }
 
   //#region navigation
 
@@ -130,33 +144,6 @@ const Landing = (props: ILandingProps) => {
 
   //#endregion
 
-  //#region notification
-
-  const defaultAlert = {
-    open: false,
-    message: "",
-    severity: "success" as AlertColor
-  }
-  const [alert, setAlert] = useState(defaultAlert)
-
-  const handleCloseAlert = (event?: Event | React.SyntheticEvent<any, Event>, reason?: SnackbarCloseReason) => {
-    if (reason === 'clickaway') {
-      return
-    }
-    setAlert(a => ({ ...defaultAlert, severity: a.severity }))
-  }
-
-  const showNotification = (message: string, severity: AlertColor = "success") => {
-    setAlert({
-      message: message,
-      severity: severity,
-      open: true
-    })
-  }
-
-  //#endregion
-
-
   //#region sorting
   const handleSorting = (event: React.MouseEvent<HTMLElement>, orderBy: string | null) => {
     if (orderBy) {
@@ -181,181 +168,10 @@ const Landing = (props: ILandingProps) => {
 
   //#region add game
 
-  interface IAddGameValidation {
-    name: string,
-    developer: string,
-    releaseDate: string,
-    summary: string,
-    genres: string,
-    platforms: string,
-    logo: string,
-    screenshots: string,
-    [index: string]: string
-  }
-
-  const [addGame, setAddGame] = useState<ICreateGame>({} as ICreateGame)
   const [addGameDialogOpen, setAddGameDialogOpen] = useState(false)
-  const [addGameDialogText, setAddGameDialogText] = useState("")
-  const [addGameErrorText, setAddGameErrorText] = useState("")
-  const [addGameValidation, setAddGameValidation] = useState<IAddGameValidation>({
-    name: "",
-    developer: "",
-    releaseDate: "",
-    summary: "",
-    genres: "",
-    platforms: "",
-    logo: "",
-    screenshots: "",
-  })
 
   const handleAddGameDialogOpen = () => setAddGameDialogOpen(true)
   const handleAddGameDialogClose = () => setAddGameDialogOpen(false)
-
-  const handleAddGameFieldChange = (e: React.ChangeEvent<HTMLTextAreaElement | HTMLInputElement>, field: string) => {
-    const value = e.target.value?.trimStart()
-    setAddGame(g => ({ ...g, [field]: value }))
-    // remove error if there is a value now
-    if (addGameValidation[field]?.length > 0 && value?.length > 0) {
-      setAddGameValidation(v => ({ ...v, [field]: '' }))
-    }
-  }
-
-  const validateAddGameForm = (): boolean => {
-    let valid = true
-    if (!addGame.name) {
-      setAddGameValidation(v => ({ ...v, name: 'Name is required' }))
-      valid = false
-    }
-
-    if (!addGame.developer) {
-      setAddGameValidation(v => ({ ...v, developer: 'Developer is required' }))
-      valid = false
-    }
-
-    if (!addGame.releaseDate) {
-      setAddGameValidation(v => ({ ...v, releaseDate: 'Release date is required' }))
-      valid = false
-    } else if (!moment(addGame.releaseDate).isValid()) {
-      setAddGameValidation(v => ({ ...v, releaseDate: 'Invalid date or format. Should be YYYY-MM-DD' }))
-      valid = false
-    }
-
-    if (!addGame.summary) {
-      setAddGameValidation(v => ({ ...v, summary: 'Summary is required' }))
-      valid = false
-    }
-
-    if (!addGame.genresIds?.length) {
-      setAddGameValidation(v => ({ ...v, genres: 'At least one genre is required' }))
-      valid = false
-    }
-
-    if (!addGame.platformsIds?.length) {
-      setAddGameValidation(v => ({ ...v, platforms: 'At least one platform is required' }))
-      valid = false
-    }
-
-    if (!addGame.logoUrl) {
-      setAddGameValidation(v => ({ ...v, logo: 'Cover is required' }))
-      valid = false
-    }
-
-    if (!addGame.screenshots?.length) {
-      setAddGameValidation(v => ({ ...v, screenshots: 'Screenshots are required' }))
-      valid = false
-    }
-
-    return valid
-  }
-
-  const handleAddGame = async () => {
-    setAddGameErrorText("")
-    if (!validateAddGameForm()) {
-      return
-    }
-
-    const newGame: ICreateGame = {
-      ...addGame,
-      releaseDate: moment(addGame.releaseDate).format("yyyy-MM-DD")
-    }
-
-    const [resp, err] = await createGame(newGame)
-    if (err) {
-      if (typeof err === 'string') {
-        setAddGameErrorText(err)
-        return
-      }
-      const error = err as IValidationResponse
-      setAddGameErrorText(error.fields?.map(f => `${f.field}: ${f.error}`).join("; ") || error.error)
-      return
-    }
-    const game = resp as IGameResponse
-    if (game.id) {
-      setAddGameDialogText("Game has been successfully added")
-      setAddGame({} as ICreateGame)
-      setTimeout(() => {
-        handleAddGameDialogClose()
-        setAddGameDialogText("")
-      }, 500)
-    } else {
-      setAddGameErrorText("An error occured. Try again later")
-    }
-  }
-
-  //#endregion
-
-  //#region upload widget
-
-  const handleLogoChanged = (fileInfo: FileInfo) => {
-    if (fileInfo.isStored) {
-      setAddGame(g => ({ ...g, logoUrl: fileInfo.cdnUrl || undefined }))
-    }
-
-    if (addGameValidation.logo.length > 0 && fileInfo.cdnUrl) {
-      setAddGameValidation(v => ({ ...v, logo: '' }))
-    }
-  }
-
-  const handleScreenshotsChanged = async (fileInfo: FileUpload | FilesUpload | null) => {
-    if (!fileInfo) {
-      return
-    }
-    let group = fileInfo as FilesUpload
-    const files = await Promise.all(group.files())
-    const urls = files.map((file) => file?.cdnUrl || '')
-    
-    setAddGame(g => ({ ...g, screenshots: urls }))
-
-    if (addGameValidation.screenshots.length > 0 && urls?.length > 0) {
-      setAddGameValidation(v => ({ ...v, screenshots: '' }))
-    }
-  }
-
-  const fileSizeLimit = (sizeInKb: number) => {
-    return (fileInfo: FileInfo) => {
-      if (fileInfo.name === null || fileInfo.size === null) {
-        return
-      }
-
-      if (fileInfo.size > sizeInKb * 1024) {
-        throw new Error('size')
-      }
-    }
-  }
-
-  const hasExtension = () => {
-    return (fileInfo: FileInfo) => {
-      if (fileInfo.name === null) {
-        return
-      }
-
-      if (!fileInfo.name.includes(".")) {
-        throw new Error('image')
-      }
-    }
-  }
-
-  const uploadValidators = [fileSizeLimit(150), hasExtension()]
 
   //#endregion
 
@@ -377,6 +193,28 @@ const Landing = (props: ILandingProps) => {
     })
   }
 
+  const handleCategoryChange = (type: "genre" | "publisher" | "developer", id: number) => {
+    if (id) {
+      if (id === navigation[type]) {
+        id = 0
+      }
+      setNavigation(p => ({
+        ...p,
+        page: defaultNavigation.page,
+        [type]: id
+      }))
+    }
+    setSearchParams(p => {
+      p.delete(pageParam)
+      if (id && id !== navigation[type]) {
+        p.set(type, id.toString())
+      } else {
+        p.delete(type)
+      }
+      return p
+    })
+  }
+
   const handleNavigation = (page: number = defaultNavigation.page) => {
     setNavigation(p => ({
       ...p,
@@ -388,15 +226,6 @@ const Landing = (props: ILandingProps) => {
     })
   }
 
-  const notifyError = (err: string | IValidationResponse | null) => {
-    if (typeof err === 'string') {
-      showNotification(err, "error")
-    } else {
-      const error = err as IValidationResponse
-      showNotification(error.fields?.map(f => `${f.field}: ${f.error}`).join("; ") || error.error, "error")
-    }
-  }
-
   // get user ratings
   useEffect(() => {
     const getRatings = async () => {
@@ -406,7 +235,7 @@ const Landing = (props: ILandingProps) => {
         gameIds: gameIds
       })
       if (err) {
-        notifyError(err)
+        setAlert(err)
         setIsLoading(false)
         return
       }
@@ -432,7 +261,7 @@ const Landing = (props: ILandingProps) => {
       } as IGamesFilter
       const [resp, err] = await fetchGames(filter, navigation.pageSize, navigation.page)
       if (err) {
-        notifyError(err)
+        setAlert(err)
         setIsLoading(false)
         return
       }
@@ -456,32 +285,46 @@ const Landing = (props: ILandingProps) => {
     getData()
   }, [navigation.page, navigation.orderBy, navigation.searchText, navigation.genre, navigation.developer, navigation.publisher])
 
-  // fetch genres
+  // fetch top genres
   useEffect(() => {
-    const getGenres = async () => {
-      const [resp, err] = await fetchGenres()
+    const getTopGenres = async () => {
+      const [resp, err] = await fetchTopGenres()
       if (err) {
-        notifyError(err)
+        setAlert(err)
         return
       }
       const genres = resp as IGenre[]
-      setGenres(genres)
+      setTopGenres(genres.slice(0, topCategoriesLimit))
     }
-    getGenres()
+    getTopGenres()
   }, [])
 
-  // fetch platforms
+  // fetch top developers
   useEffect(() => {
-    const getPlatforms = async () => {
-      const [resp, err] = await fetchPlatforms()
+    const getTopDevelopers = async () => {
+      const [resp, err] = await fetchTopCompanies("dev")
       if (err) {
-        notifyError(err)
+        setAlert(err)
         return
       }
-      const platforms = resp as IPlatform[]
-      setPlatforms(platforms)
+      const developers = resp as ICompany[]
+      setTopDevelopers(developers.slice(0, topCategoriesLimit))
     }
-    getPlatforms()
+    getTopDevelopers()
+  }, [])
+
+  // fetch top publishers
+  useEffect(() => {
+    const getTopPublishers = async () => {
+      const [resp, err] = await fetchTopCompanies("pub")
+      if (err) {
+        setAlert(err)
+        return
+      }
+      const publishers = resp as ICompany[]
+      setTopPublishers(publishers.slice(0, topCategoriesLimit))
+    }
+    getTopPublishers()
   }, [])
 
   const searchFieldProps: ISearchFieldProps = {
@@ -489,7 +332,9 @@ const Landing = (props: ILandingProps) => {
     changeText: handleSearchTextChange
   }
 
-  const uwpk = '8869032692db5708aebb'
+  const categoryApplied = (): boolean => {
+    return !!navigation.genre || !!navigation.developer || !!navigation.publisher
+  }
 
   return (
     <LocalizationProvider dateAdapter={DateAdapter}>
@@ -502,221 +347,19 @@ const Landing = (props: ILandingProps) => {
             <CircularProgress color="inherit" />
           </Backdrop>
 
-          <Modal
-            fullwidth={matchesMd}
-            matchesMd={matchesMd}
-            isOpen={addGameDialogOpen}
-            closeDialog={handleAddGameDialogClose}
-            title='Add new game'
-            dialogText={addGameDialogText}
-            dialogErrorText={addGameErrorText}
-            submitActionName='Add game'
-            handleSubmit={handleAddGame}
-          >
-            <>
-              <Grid item sx={{ minWidth: matchesMd ? fieldWidthLarge : fieldWidthSmall }}>
-                <TextField
-                  required
-                  error={!!addGameValidation.name}
-                  helperText={addGameValidation.name}
-                  fullWidth
-                  label="Name"
-                  margin="normal"
-                  value={addGame?.name || ""}
-                  onChange={(e: React.ChangeEvent<HTMLTextAreaElement | HTMLInputElement>) => handleAddGameFieldChange(e, 'name')}
-                />
-              </Grid>
-              <Grid item sx={{ minWidth: matchesMd ? fieldWidthLarge : fieldWidthSmall }}>
-                <TextField
-                  required
-                  error={!!addGameValidation.developer}
-                  helperText={addGameValidation.developer}
-                  fullWidth
-                  label="Developer"
-                  margin="normal"
-                  value={addGame?.developer || ""}
-                  onChange={(e: React.ChangeEvent<HTMLTextAreaElement | HTMLInputElement>) => handleAddGameFieldChange(e, 'developer')}
-                />
-              </Grid>
-              <Grid item sx={{ minWidth: matchesMd ? fieldWidthLarge : fieldWidthSmall }}>
-                {isTouchDevice()
-                  ? <MobileDatePicker
-                    label="Release date"
-                    inputFormat="yyyy-MM-DD"
-                    value={addGame?.releaseDate || null}
-                    onChange={(d: string | null) => {
-                      setAddGameValidation(v => ({ ...v, releaseDate: "" }))
-                      setAddGame(g => ({ ...g, releaseDate: d || "" }))
-                    }}
-                    mask="____-__-__"
-                    renderInput={(params: TextFieldProps) =>
-                      <TextField {...params}
-                        fullWidth
-                        margin="normal"
-                        required
-                        error={!!addGameValidation.releaseDate}
-                        helperText={addGameValidation.releaseDate}
-                      />}
-                  />
-                  : <DesktopDatePicker
-                    label="Release date"
-                    inputFormat="yyyy-MM-DD"
-                    value={addGame?.releaseDate || null}
-                    onChange={(d: string | null) => {
-                      setAddGameValidation(v => ({ ...v, releaseDate: "" }))
-                      setAddGame(g => ({ ...g, releaseDate: d || "" }))
-                    }}
-                    mask="____-__-__"
-                    renderInput={(params: TextFieldProps) =>
-                      <TextField {...params}
-                        fullWidth
-                        margin="normal"
-                        required
-                        error={!!addGameValidation.releaseDate}
-                        helperText={addGameValidation.releaseDate}
-                      />}
-                  />
-                }
-              </Grid>
-              <Grid item sx={{ minWidth: matchesMd ? fieldWidthLarge : fieldWidthSmall }}>
-                <TextField
-                  required
-                  error={!!addGameValidation.summary}
-                  helperText={addGameValidation.summary}
-                  fullWidth
-                  multiline
-                  margin="normal"
-                  label="Summary"
-                  value={addGame?.summary || ""}
-                  onChange={(e: React.ChangeEvent<HTMLTextAreaElement | HTMLInputElement>) => handleAddGameFieldChange(e, "summary")}
-                />
-              </Grid>
-              <Grid item sx={{ minWidth: matchesMd ? fieldWidthLarge : fieldWidthSmall }}>
-                <TextField
-                  fullWidth
-                  margin="normal"
-                  label="Websites"
-                  placeholder="mygame.com,twitch.com/mygame"
-                  value={addGame?.websites?.join(",") || ""}
-                  onChange={(e: React.ChangeEvent<HTMLTextAreaElement | HTMLInputElement>) => setAddGame(g => ({
-                    ...g,
-                    websites: e.target.value?.split(",").map(g => g.trim())
-                  }))}
-                />
-              </Grid>
-              <Grid item sx={{ minWidth: matchesMd ? fieldWidthLarge : fieldWidthSmall }}>
-              <Autocomplete
-                multiple
-                disableCloseOnSelect
-                filterSelectedOptions
-                id="genres"
-                options={genres}
-                getOptionLabel={(option: IGenre) => option.name}
-                onChange={(e: React.SyntheticEvent, value: IGenre[], reason: AutocompleteChangeReason, details?: AutocompleteChangeDetails<IGenre> | undefined) => {
-                  setAddGame(g => ({
-                    ...g,
-                    genresIds: value?.map(g => g.id)
-                  }))
+          <AddGameModal
+            handleAddGameDialogClose={handleAddGameDialogClose}
+            addGameDialogOpen={addGameDialogOpen}
+          />
 
-                  if (addGameValidation.genres?.length !== 0 && value?.length > 0) {
-                    setAddGameValidation(v => ({ ...v, genres: '' }))
-                  }
-                }}
-                renderInput={(params) => (
-                  <TextField
-                    {...params}
-                    required
-                    fullWidth
-                    error={!!addGameValidation.genres}
-                    helperText={addGameValidation.genres}
-                    margin="normal"
-                    label="Genres"
-                  />
-                )}
-              />
-              <Autocomplete
-                multiple
-                disableCloseOnSelect
-                filterSelectedOptions
-                id="platforms"
-                options={platforms}
-                getOptionLabel={(option: IPlatform) => option.name}
-                onChange={(e: React.SyntheticEvent, value: IPlatform[], reason: AutocompleteChangeReason, details?: AutocompleteChangeDetails<IPlatform> | undefined) => {
-                  setAddGame(g => ({
-                    ...g,
-                    platformsIds: value?.map(p => p.id)
-                  }))
-
-                  if (addGameValidation.platforms?.length !== 0 && value?.length > 0) {
-                    setAddGameValidation(v => ({ ...v, platforms: '' }))
-                  }
-                }}
-                renderInput={(params) => (
-                  <TextField
-                    {...params}
-                    required
-                    fullWidth
-                    error={!!addGameValidation.platforms}
-                    helperText={addGameValidation.platforms}
-                    margin="normal"
-                    label="Platforms"
-                  />
-                )}
-              />
-              </Grid>
-              <Grid item sx={{ minWidth: matchesMd ? fieldWidthLarge : fieldWidthSmall }}>
-                <label htmlFor={'coverUploadWidget'}>Cover </label>
-                <Typography variant="caption" color={"rgba(200, 150, 100, 0.8)"}> (max size 150 kb, ratio 3:4)</Typography>
-                <div id={'coverUploadWidget'}>
-                  <UploadWidget
-                    imagesOnly
-                    previewStep
-                    clearable
-                    crop='3:4'
-                    tabs='file'
-                    publicKey={uwpk}
-                    validators={uploadValidators}
-                    onChange={(fileInfo: FileInfo) => handleLogoChanged(fileInfo)}
-                  />
-                </div>
-                <Typography variant="caption" color="#f44336">  {addGameValidation.logo}</Typography>
-              </Grid>
-              <Grid item sx={{ minWidth: matchesMd ? fieldWidthLarge : fieldWidthSmall }}>
-                <label htmlFor={'screenshotsUploadWidget'}>Screenshots </label>
-                <Typography variant="caption" color={"rgba(200, 150, 100, 0.8)"}> (max size 150 kb, ratio 9:5, max 7 images)</Typography>
-                <div id={'screenshotsUploadWidget'}>
-                  <UploadWidget
-                    imagesOnly
-                    multiple
-                    previewStep
-                    clearable
-                    crop='9:5'
-                    multipleMax={7}
-                    tabs='file'
-                    publicKey={uwpk}
-                    validators={uploadValidators}
-                    onFileSelect={(fileInfo: FileUpload | FilesUpload | null) => handleScreenshotsChanged(fileInfo)}
-                  />
-                </div>
-                <Typography variant="caption" color="#f44336">  {addGameValidation.screenshots}</Typography>
-              </Grid>
-            </>
-          </Modal>
-
-          <Snackbar
-            anchorOrigin={{ vertical: 'top', horizontal: 'right' }}
-            open={alert.open}
-            autoHideDuration={5000}
-            onClose={handleCloseAlert}
-          >
-            <Alert onClose={handleCloseAlert} severity={alert.severity} sx={{ width: '100%' }} >
-              {alert.message}
-            </Alert>
-          </Snackbar>
+          <Notification 
+            message={alert} 
+            resetMessage={() => setAlert(null)}
+          />
 
           <Box sx={{ pb: 3 }}>
-            <Grid container spacing={2} direction="row" justifyContent="space-between" alignItems="center" sx={{ pb: 3 }}>
-              <Grid item xs={3} sx={{ pt: 0 }}>
+            <Grid container spacing={2} direction="row" justifyContent="space-between" alignItems="center" sx={{ pb: topCategoriesOpen ? 0.5 : 3 }}>
+              <Grid item xs={matchesXs ? 4 : 3} sx={{ pt: 0 }}>
                 <ToggleButtonGroup
                   value={navigation.orderBy}
                   size="small"
@@ -725,28 +368,144 @@ const Landing = (props: ILandingProps) => {
                   aria-label="sorting"
                 >
                   <ToggleButton value="default" aria-label="default" title="Ranking">
-                    <WhatshotIcon fontSize={matchesXs ? "small" : "medium"} />
+                    <WhatshotIcon fontSize={mediaQueryToTextSize()} />
                   </ToggleButton>
                   <ToggleButton value="releaseDate" aria-label="release date" title="Release date">
-                    <DateRangeIcon fontSize={matchesXs ? "small" : "medium"} />
+                    <DateRangeIcon fontSize={mediaQueryToTextSize()} />
                   </ToggleButton>
                   <ToggleButton value="name" aria-label="name" title="Name">
-                    <AbcIcon fontSize={matchesXs ? "small" : "medium"} />
+                    <AbcIcon fontSize={mediaQueryToTextSize()} />
                   </ToggleButton>
                 </ToggleButtonGroup>
               </Grid>
-              <Grid item sm={6} sx={{ textAlign: "center" }}>
+              <Grid item xs={3} sx={{ textAlign: "right" }}>
                 <Typography variant={matchesXs ? "h6" : "h5"}>
                   Games <sup style={{ fontSize: matchesXs ? 9 : 11, color: "" }}> {count}</sup>
                 </Typography>
               </Grid>
-              <Grid item xs={3} sx={{ textAlign: "right" }}>
+              <Grid item xs={3} sx={{ textAlign: "left" }}>
                 {hasRole([roles.publisher])
-                  ? <Button variant="contained" onClick={() => handleAddGameDialogOpen()}>Add game</Button>
+                  ? <Button variant="contained" size={mediaQueryToSize()} onClick={() => handleAddGameDialogOpen()}>Add game</Button>
                   : <Box sx={{ width: "25w" }} />
                 }
               </Grid>
+              <Grid item xs={2} sx={{ textAlign: "right" }}>
+                <ToggleButton
+                  value="check"
+                  selected={topCategoriesOpen}
+                  onChange={() => {setTopCategoriesOpen(v => !v) }}
+                  size="small"
+                  title="Categories"
+                >
+                  <CategoryOutlinedIcon fontSize={mediaQueryToTextSize()} sx={{ color: categoryApplied() && selectedCategoryColor || "" }} />
+                  {topCategoriesOpen 
+                  ?  <KeyboardArrowUpOutlinedIcon fontSize={mediaQueryToTextSize()} sx={{ color: categoryApplied() && selectedCategoryColor || "" }} />
+                  : <KeyboardArrowDownOutlinedIcon fontSize={mediaQueryToTextSize()} sx={{ color: categoryApplied() && selectedCategoryColor || "" }} />
+                  }
+                </ToggleButton>
+              </Grid>
             </Grid>
+
+            { topCategoriesOpen &&
+              <Grid container spacing={1} direction="row" justifyContent="center" alignItems="flex-start" sx={{ pb: 2, pt: 0.5 }}>
+                <Grid item xs={4} sx={{ textAlign: "center", pl: 1 }} >
+                  <>
+                  <Chip
+                    label={<Typography color="primary" variant="body1">Genres</Typography>}
+                    variant="outlined" 
+                    color="info" 
+                  />
+                  </>
+                </Grid>
+                <Grid item xs={4} sx={{ textAlign: "center" }} >
+                  <Chip
+                    label={<Typography color="primary" variant="body1">Publishers</Typography>}
+                    variant="outlined" 
+                    color="info" 
+                  />
+                </Grid>
+                <Grid item xs={4} sx={{ textAlign: "center" }} >
+                  <Chip
+                    label={<Typography color="primary" variant="body1">Developers</Typography>}
+                    variant="outlined" 
+                    color="info" 
+                  />
+                </Grid>
+
+                <Grid item xs={4} sx={{ pt: 0, pl: 0 }} >
+                  <List sx={{ pt: 0 }} >
+                    {topGenres.map((genre: IGenre) => (
+                      <ListItem key={genre.id} dense>
+                        <ListItemButton 
+                          component="a"
+                          dense
+                          selected={genre.id === navigation.genre}
+                          disableGutters={matchesXs}
+                          onClick={(event: React.MouseEvent<HTMLAnchorElement, MouseEvent>) => {
+                            event.preventDefault() // Prevent default anchor link behavior
+                            handleCategoryChange("genre", genre.id)
+                          }}
+                        >
+                          <ListItemText 
+                            primary={genre.name}
+                            primaryTypographyProps={{ textAlign:"center", flex: 1, color: genre.id === navigation.genre ? selectedCategoryColor : "" }}
+                            sx={{ mt: 0, mb: 0 }} 
+                          />
+                        </ListItemButton>
+                      </ListItem>
+                    ))}
+                  </List>
+                </Grid>
+                <Grid item xs={4} sx={{ pt: 0 }} >
+                  <List sx={{ pt: 0 }} >
+                    {topPublishers.map((publisher: ICompany) => (
+                      <ListItem key={publisher.id} dense>
+                        <ListItemButton 
+                          component="a"
+                          dense
+                          selected={publisher.id === navigation.publisher}
+                          disableGutters={matchesXs}
+                          onClick={(event: React.MouseEvent<HTMLAnchorElement, MouseEvent>) => {
+                            event.preventDefault() // Prevent default anchor link behavior
+                            handleCategoryChange("publisher", publisher.id)
+                          }}
+                        >
+                          <ListItemText 
+                            primary={publisher.name}
+                            primaryTypographyProps={{ textAlign:"center", flex: 1, color: publisher.id === navigation.publisher ? selectedCategoryColor : "" }}
+                            sx={{ mt: 0, mb: 0 }} 
+                          />
+                        </ListItemButton>
+                      </ListItem>
+                    ))}
+                  </List>
+                </Grid>
+                <Grid item xs={4} sx={{ pt: 0 }} >
+                  <List sx={{ pt: 0 }} >
+                    {topDevelopers.map((developer: ICompany) => (
+                      <ListItem key={developer.id} dense>
+                        <ListItemButton 
+                          component="a"
+                          dense
+                          selected={developer.id === navigation.developer}
+                          disableGutters={matchesXs}
+                          onClick={(event: React.MouseEvent<HTMLAnchorElement, MouseEvent>) => {
+                            event.preventDefault() // Prevent default anchor link behavior
+                            handleCategoryChange("developer", developer.id)
+                          }}
+                        >
+                          <ListItemText 
+                            primary={developer.name} 
+                            primaryTypographyProps={{ textAlign:"center", flex: 1, color: developer.id === navigation.developer ? selectedCategoryColor : "" }}
+                            sx={{ mt: 0, mb: 0 }}
+                          />
+                        </ListItemButton>
+                      </ListItem>
+                    ))}
+                  </List>
+                </Grid>
+              </Grid>
+            }
 
             <GameDetails
               game={selectedGame}
